refactor: Restructure the models folder to seperate database models and web models
