+++ src/models/__init__.py
@@ -1,7 +1,1 @@
-"""Models Initialization Module."""
-
-from src.models.animap import AniMap
-from src.models.base import Base
-from src.models.housekeeping import Housekeeping
-
-__all__ = ["Base", "AniMap", "Housekeeping"]
+"""PlexAniBridge application models."""
