+++ src/models/db/__init__.py
@@ -1,0 +1,8 @@
+"""Models for PlexAniBridge database tables."""
+
+from src.models.db.animap import AniMap
+from src.models.db.base import Base
+from src.models.db.housekeeping import Housekeeping
+from src.models.db.sync_history import SyncHistory
+
+__all__ = ["Base", "AniMap", "Housekeeping", "SyncHistory"]
