+++ src/models/schemas/__init__.py
@@ -1,0 +1,1 @@
+"""Schemas for external APIs used in PlexAniBridge."""
