+++ src/models/schemas/anilist.py
